requirements.txt
streamlit
pandas
requests
pymupdf
python-docx
openpyxl
python-pptx
openpyxl
google-auth
google-auth-oauthlib
google-api-python-client
